+++ external-crates/move/crates/move-model-2/src/compiled_model.rs
@@ -1,7 +1,14 @@
 // Copyright (c) The Move Contributors
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::model::{self, WithoutSource};
+use crate::{
+    model::{self, PackageData, WithoutSource},
+    normalized, serializable_signatures,
+};
+use move_binary_format::CompiledModule;
+use move_core_types::account_address::AccountAddress;
+use move_symbol_pool::Symbol;
+use std::{cell::OnceCell, collections::BTreeMap};
 
 pub type Model = model::Model<WithoutSource>;
 pub type Package<'a> = model::Package<'a, WithoutSource>;
@@ -12,3 +19,77 @@ pub type Struct<'a> = model::Struct<'a, WithoutSource>;
 pub type Enum<'a> = model::Enum<'a, WithoutSource>;
 pub type Variant<'a> = model::Variant<'a, WithoutSource>;
 pub type Function<'a> = model::Function<'a, WithoutSource>;
+pub type CompiledConstant<'a> = model::CompiledConstant<'a, WithoutSource>;
+
+impl Model {
+    pub fn from_compiled(
+        named_address_reverse_map: &BTreeMap<AccountAddress, Symbol>,
+        modules: Vec<CompiledModule>,
+    ) -> Self {
+        let compiled = normalized::Packages::new(&modules);
+        let packages = compiled
+            .packages
+            .values()
+            .map(|package| {
+                let addr = package.package;
+                let data = PackageData::from_compiled(named_address_reverse_map, package);
+                (addr, data)
+            })
+            .collect();
+        let root_named_address_map = named_address_reverse_map
+            .iter()
+            .map(|(a, n)| (*n, *a))
+            .collect();
+        let mut model = Self {
+            files: None,
+            root_package_name: None,
+            root_named_address_map,
+            info: None,
+            compiled,
+            packages,
+            serializable_signatures: OnceCell::new(),
+            _phantom: std::marker::PhantomData,
+        };
+        model.compute_dependencies();
+        model.compute_function_dependencies();
+        model.check_invariants();
+        model
+    }
+
+    pub fn serializable_signatures(&self) -> &serializable_signatures::Packages {
+        self.serializable_signatures
+            .get_or_init(|| serializable_signatures::Packages::from(&self.compiled))
+    }
+}
+
+impl<'a> Module<'a> {
+    pub fn maybe_member(&self, name: impl Into<Symbol>) -> Option<Member<'a>> {
+        let name = name.into();
+        self.maybe_struct(name)
+            .map(Member::Struct)
+            .or_else(|| self.maybe_enum(name).map(Member::Enum))
+            .or_else(|| self.maybe_function(name).map(Member::Function))
+    }
+
+    pub fn member(&self, name: impl Into<Symbol>) -> Member<'a> {
+        self.maybe_member(name).unwrap()
+    }
+
+    pub fn constants(&self) -> impl Iterator<Item = CompiledConstant<'a>> + '_ {
+        self.compiled
+            .constants
+            .iter()
+            .enumerate()
+            .map(|(idx, compiled)| CompiledConstant {
+                module: *self,
+                compiled,
+                data: &self.data.constants[idx],
+            })
+    }
+}
+
+impl<'a> Function<'a> {
+    pub fn compiled(&self) -> &'a normalized::Function {
+        self.compiled.unwrap()
+    }
+}
